requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
black==23.3.0
charset-normalizer==3.1.0
click==8.1.3
cogwatch==3.3.1
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
discord.py==2.3.2
dpytest==0.7.0
exceptiongroup==1.2.0
fonttools==4.39.3
frozenlist==1.3.3
idna==3.4
iniconfig==2.0.0
kaleido==0.2.1
kiwisolver==1.4.4
matplotlib==3.7.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
packaging==23.0
pandas==2.2.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
plotly==5.20.0
pluggy==1.4.0
pyparsing==3.0.9
pytest==8.1.1
pytest-asyncio==0.23.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
six==1.16.0
sniffio==1.3.0
tenacity==8.2.3
tomli==2.0.1
tzdata==2024.1
watchfiles==0.15.0
yarl==1.8.2
